# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Blog Project CRUD",
    "description": "My Blog Project Application API",
    "license": {
      "name": "Oleksii Umanets",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:5000",
  "basePath": "/",
  "tags": [
    {
      "name": "Blog",
      "description": "API for blog app"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["User"],
        "summary": "Create new user",
        "description": "",
        "operationId": "create user",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "description": "Sign in user",
            "required": false,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "Sign in user",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "Sign in user",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid data supplied"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["User"],
        "summary": "Sign in user to the system",
        "description": "",
        "operationId": "createSession",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "description": "Log in user",
            "required": false,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "Log in user",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "Log in user",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid data supplied"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/auth/logout": {
      "delete": {
        "tags": ["User"],
        "summary": "Deletes session",
        "description": "",
        "operationId": "deleteSession",
        "produces": ["application/xml", "application/json"],
        "responses": {
          "422": {
            "description": "Unprocessable entity"
          }
        }
      }
    },
    "/api/auth/authchecker": {
      "get": {
        "tags": ["User"],
        "summary": "Check if user is signed in",
        "description": "Check if user is signed in",
        "operationId": "checkAuth",
        "produces": ["application/xml", "application/json"],
        "responses": {
          "200": {
            "description": "Authenticated Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/home/all-users": {
      "get": {
        "tags": ["User"],
        "summary": "Get all users",
        "description": "Get all users",
        "operationId": "allUsers",
        "produces": ["application/xml", "application/json"],
        "responses": {
          "200": {
            "description": "Get all user successfully"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/post/create": {
      "get": {
        "tags": ["Post"],
        "summary": "Create new post",
        "description": "Create new post",
        "operationId": "createPost",
        "produces": ["application/xml", "application/json"],
        "responses": {
          "200": {
            "description": "Post was successfully created"
          },
          "400": {
            "description": "Image required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/post/{id}": {
      "get": {
        "tags": ["Post"],
        "summary": "Get post",
        "description": "Get post",
        "operationId": "getPost",
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of post that we want to get",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Post was successfully received"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/post/{postId}/like": {
      "put": {
        "tags": ["Post"],
        "summary": "Like or dislike post",
        "description": "Like or dislike post",
        "operationId": "likePost",
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "postId of post that we want to like/dislike",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Post was successfully liked"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/post/{postId}": {
      "delete": {
        "tags": ["Post"],
        "summary": "Delete post",
        "description": "Delete post",
        "operationId": "deletePost",
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "postId of post that we want to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Post was successfully deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      },

      "put": {
        "tags": ["Post"],
        "summary": "Edit post",
        "description": "Edit post",
        "operationId": "editPost",
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "postId of post that we want to edit",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Post was successfully edited"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/settings/profile": {
      "put": {
        "tags": ["User"],
        "summary": "Change profile info",
        "description": "",
        "operationId": "changeProfileInfo",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "description": "Change user info",
            "required": false,
            "type": "string"
          },
          {
            "name": "profileImg",
            "in": "formData",
            "description": "Change user info",
            "required": false,
            "type": "string"
          },
          {
            "name": "bio",
            "in": "formData",
            "description": "Change user info",
            "required": false,
            "type": "string"
          },
          {
            "name": "birthday",
            "in": "formData",
            "description": "Change user info",
            "required": false,
            "type": "string"
          },
          {
            "name": "gender",
            "in": "formData",
            "description": "Change user info",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid data supplied"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "tags": ["User"],
        "summary": "Get profile info",
        "description": "",
        "operationId": "getProfileInfo",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/xml", "application/json"],
        "responses": {
          "400": {
            "description": "Invalid data supplied"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/user/{username}/info": {
      "get": {
        "tags": ["User"],
        "summary": "Get profile page info",
        "description": "",
        "operationId": "getProfilePageInfo",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "username of user which info we want to get",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid data supplied"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/user/{username}/posts": {
      "get": {
        "tags": ["Post"],
        "summary": "Get profile page posts",
        "description": "",
        "operationId": "getProfilePagePosts",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "username of user which posts we need to get",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid data supplied"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/user/{username}/follow": {
      "post": {
        "tags": ["User"],
        "summary": "Follow/Unfollow user",
        "description": "",
        "operationId": "Follow/Unfollow user",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "username of user that needs to follow/unfollow",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid data supplied"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/user/{username}/followers": {
      "get": {
        "tags": ["User"],
        "summary": "Get user followers",
        "description": "",
        "operationId": "Get user followers",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "username of user that needs to get followers",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid data supplied"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/user/{username}/following": {
      "get": {
        "tags": ["User"],
        "summary": "Get user following",
        "description": "",
        "operationId": "Get user following",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "username of user that needs to get following",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid data supplied"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "required": ["email", "username", "password"],
      "properties": {
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "username": {
          "type": "string",
          "uniqueItems": true
        },
        "password": {
          "type": "string",
          "uniqueItems": true
        },
        "posts": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "likes": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "birthday": {
          "type": "date"
        },
        "profileImg": {
          "type": "string"
        },
        "profileImg_public_id": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "followers": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "following": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "Post": {
      "required": ["image", "author"],
      "properties": {
        "image": {
          "type": "string",
          "uniqueItems": true
        },

        "author": {
          "type": "string",
          "uniqueItems": true
        },
        "likes": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "date": {
          "type": "date"
        },
        "text": {
          "type": "string"
        },
        "public_id": {
          "type": "string"
        },
        "hashtags": {
          "type": "string"
        }
      }
    }
  }
}